requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
stripe>=7.0.0
requests>=2.31.0
django-redis>=5.4.0
drf-spectacular>=0.27.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
redis>=5.0.0
Pillow>=10.0.0
gunicorn>=21.2.0
